SUCCESS!!! Parallax works.
Parallax works.

Had to create a .wrapper:

.welcome-page-wrapper{
   height: 100vh;
   overflow-x: hidden;
   overflow-y: auto;
   perspective: 10px;
}

Inside this .wrapper is where I will put the components (inner components) that will be involved in the parallax scroll.

The css logic will be inside each inner component and would involve using the translateZ( ) to move the image in a 3D perspective relative to the perspective in wrapper and then scaling with scale( ) css prop..

Also note the preserve 3d:

transform-style: preserve-3d;

Inside the <Header> which contains the back video, front image and logo.