Used the use hook to create route founction.
WHY DID I USE THE use Hook?

Because inside that use hook, I plan to use the useLocation hook to filter out routes that are NOT my current page.

So, if I am currently on the Home page, my route path would be "/home'.

I want to use location to find all routes inside either generalRoutes or loggedInRoutes function to return routes that are NOT '/home'.

WHY???

I plan to use those inside the navigation bars (found inside navigationbars).

How do I plan to  use them inside the navigation bars?

I plan to create links to pages that are NOT the pages I am on.

Again, if I am on the home page, my general navigation bar would have the following links:

/about
/contact
/any other link that is NOT home.

I mean, it doesn't make sense to have a link that leads to a parge you are currently on.