Create Registration.jsx
This will be a card inside the registration page that, when clicked, will trigger a Modal (that I have not created yet).

So, when the user goes to the Registration page, he will see:

LOGIN CARD                                 SIGNUP CARD

Each card is "clickable" (please see the onClick embedded in the card).

You click it and the modal (that I have not created yet) pops up. Actually, the modal toggles.

I MAY need a useState( ), but I'm not sure.