Made .header: hidden... BUT may change.
So, if you look at the green highlight under .header { ... }, you will see overflow: hidden.

It is highlighted in GREEN.

The reason I made it hidden FOR NOW is because the video is "flowing" to the bottom of the header and now, you can see the <video> (.back-video) bleed out to the bottom.

If I make it hidden, it solves the video being seen below, but it seems the parallax effect is now gone???

I may need to address this later.