Changed setState for Input inviteOthers = no.
If the user does NOT want to invite others to the challenge, the entire challengeAnnouncement state is now set to what it was in the beginning:

setChallengeAnnouncement(prv => ({...prv, id: '', _announcementOwnerID: '', headline: '', description: '', announcementEndsOn: '0000-00-00', _challengeAnnouncementID: '' })

WHY?

Because, what if the user enters "yes" (invite others), starts to fill out the form and then midway, the user changes his mind and hits "no"? The component for that form will go away... but whatever the user already entered stays in the state.