Update AddVideosToChallenge.jsx
Added a useEffect with a dependency of actualChallenge.challengeVideoIDs.length that will call showEligibleVideos( ) every time the actualChallenge.challengeVideoIDs.length changes.

By doing this, whenever I add a video (after clicking 'JOIN THIS CHALLENGE' and selecting a video), the state updates and the video would move to VIDEOS CURRENTLY IN CHALLENGE.

Otherwise, I have to refresh to see the change.