logic change to setGridTemplateColumns.
I changed the logic for curentChallengesRef.current.style.setGridTemplateColumns because I found out that...

When the page first renders, curentChallengesRef.current.style.setGridTemplateColumns ! = 19%. It is actually undefined (or nothing).

The "19%" is calculated after the <Container style={{ display: grid }} > and the inside ChallengeWrapper renders and by that time, it's too late.

SOLUTION: Instead of setting dependency to [ ], I set it to [ currentChallengesRef?.current?.style ].