handleExpiredActiveChallenges.
Problem: Inside fetchDataThenSetState, one of the arguments is fetchData which already returns the data as .json( ).

What I did was:

data = response.json( );

Which is where the error was: response is ALREADY a json( ). I should not just do response.json( ) because in fetchDataAPI, it is already returned as json( ) data.

I should have just setState( data );

Since fetchDataThenSetState is also inside handeExpiredActiveChallenges inside AppJsxFunctions, then that is why handleExpiredActiveChallenges threw an error.